Add files via upload
STEP and STL files for a hand-held drill powered scooter. The design consists of a gearing system that transfers the torque from a battery powered cordless drill to the front wheel axle of the scooter. The shaft that connects to the drill is jointed to the inner rim of a steel ball bearing.  The design requires a connection from the Eye bolt on the handle mounted lever to the trigger on the drill, either by string or wire, to facilitate motion.